refactor(download): rename files per naming guideline (#1193)
- capability.ts → capabilityUtils.ts (P2: functions-only)
- helpers/authFetch.ts → helpers/authFetchUtils.ts (P2: functions-only)
- helpers/localProxy.ts → helpers/localProxyUtils.ts (P2: functions-only)
- ports.ts → types.ts (P2: types-only aggregation)
- postJson.ts → postJsonUtils.ts (P2: functions-only)

Rollback: git revert <hash> --no-edit
