+++ packages/download/src/helpers/authFetchUtils.ts
