+++ packages/download/src/helpers/localProxyUtils.ts
