+++ packages/download/src/index.ts
@@ -2,8 +2,8 @@ export * from './DownloadService.js';
 export * from './adapters/FetchNetworkPort.js';
 export * from './authNotifications.js';
 export * from './helpers/createAuthAwareNetworkPort.js';
-export * from './helpers/authFetch.js';
-export * from './postJson.js';
-export * from './ports.js';
+export * from './helpers/authFetchUtils.js';
+export * from './postJsonUtils.js';
+export * from './types.js';
 export * from './smartFetch.js';
 export { getCorsProxyBaseURL, setCorsProxyBaseURL } from './helpers/resolveNetworkUrl.js';
